+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Upgrade the `com.google.testing.compile` artifact to version `0.18-rf`.
 * Upgrade the `org.realityforge.proton` artifacts to version `0.32`.
 * Upgrade the `org.realityforge.guiceyloops` artifact to version `0.106`.
+* Add `CssProps.borderRadius(...)` helper methods.
 
 ### [v0.148](https://github.com/react4j/react4j/tree/v0.148) (2020-01-15) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.147...v0.148)